[Feature-19-implement-security][WIP] Fix CORS problem. Fixed integration tests failing after changes. Refactored services related to security to be in their package in /service package.
