+++ src/main/java/jewellery/inventory/security/JwtAuthenticationFilter.java
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import jewellery.inventory.exception.security.jwt.JwtTokenNotFoundException;
+import jewellery.inventory.service.security.JwtTokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
