+++ src/main/java/jewellery/inventory/service/security/JwtTokenService.java
@@ -1,4 +1,4 @@
-package jewellery.inventory.security;
+package jewellery.inventory.service.security;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -10,6 +10,7 @@ import java.util.Map;
 import java.util.function.Function;
 import jewellery.inventory.exception.security.jwt.JwtExpiredException;
 import jewellery.inventory.exception.security.jwt.JwtIsNotValidException;
+import jewellery.inventory.security.JwtUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.BadCredentialsException;
