+++ src/test/java/jewellery/inventory/integration/AuthenticatedIntegrationTestBase.java
@@ -7,7 +7,7 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 
 import java.util.Collections;
 import jewellery.inventory.model.User;
-import jewellery.inventory.security.JwtTokenService;
+import jewellery.inventory.service.security.JwtTokenService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
