+++ src/test/java/jewellery/inventory/unit/service/security/AuthServiceTest.java
@@ -15,8 +15,8 @@ import jewellery.inventory.exception.security.InvalidCredentialsException;
 import jewellery.inventory.mapper.UserMapper;
 import jewellery.inventory.model.User;
 import jewellery.inventory.repository.UserRepository;
-import jewellery.inventory.security.JwtTokenService;
-import jewellery.inventory.service.AuthService;
+import jewellery.inventory.service.security.JwtTokenService;
+import jewellery.inventory.service.security.AuthService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
