+++ src/test/resources/application.properties
@@ -8,5 +8,6 @@ spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDri
 spring.jpa.show-sql=true
 cors.allowedOrigins=http://localhost:3000
 cors.allowedMethods=GET,POST,PUT,DELETE
+cors.allowedHeaders=Authorization,Cache-Control,Content-Type
 jwt.secret.key=QdzigVY4XWNItestqpRdNuCGXx+FXok5e++GeMm1OlE=
 jwt.token.expiration=#{1000 * 60 * 24}
